requirements.txt
flask
cohere
pdfminer.six
faiss-cpu
python-dotenv